Completely redone world pipeline. Summary of marjor improvements: -Modular stages -Node caching -Node abandoning -Instant zoom out -Moving nodes around, deleting nodes, and node R/W is safer with minimal locking -Multi-instance
Other noteworthy changes include minor sampler changes. Still messy but it's easier to change sample type.

+++ ProjectIW/OctreeNode.h
@@ -40,7 +40,7 @@ namespace MDC
 
 		XMINT3 position;
 		int scale;
-		short size;
+		int size;
 		unsigned char child_index;
 		bool is_leaf;
 		unsigned char corners;
